+++ c/utils.h
@@ -239,17 +239,11 @@ void  _dt_free(void *ptr);
     return NULL;                                                               \
 } while (0)
 
-#define dterra0(message) do {                                                  \
-    _dt_err_a(message "\nat %s#L%d", __FILE__, __LINE__);                      \
-    return NULL;                                                               \
-} while (0)
-
 
 // External functions -- should be defined in py_utils.c
 // These functions merely set the error message, and nothing else.
 void _dt_err_r(const char *format, ...) __attribute__ ((format(printf, 1, 2)));
 void _dt_err_v(const char *format, ...) __attribute__ ((format(printf, 1, 2)));
-void _dt_err_a(const char *format, ...) __attribute__ ((format(printf, 1, 2)));